requirements.txt

ollama
fastapi
uvicorn
python-dotenv
pyttsx3
speechrecognition
